requirements.txt
# 依赖包列表（已锁定版本，确保部署一致性）
fastapi==0.124.4
uvicorn==0.38.0
crewai==1.6.1
chromadb==1.1.1
pydantic==2.11.10
langchain-openai==1.1.10
langchain-google-genai==4.2.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
python-multipart==0.0.20
edge-tts==7.2.7
slowapi==0.1.9
